+++ htdocs/components/RequestDashboardComponent/subComponents/TimeRequestComponent.js
@@ -63,6 +63,20 @@ class TimeRequestComponent extends MainComponent {
                 }, problem.problemID)
             },
             {
+                path: "linkedSalesOrderID",
+                label: "",
+                hdToolTip: "Sales Order",
+                hdClassName: "text-center",
+                icon: "fal fa-2x fa-tag color-gray2 pointer",
+                className: "text-center",
+                classNameColumn: "",
+                sortable: false,
+                content: (problem) => el('a', {
+                    href: `SalesOrder.php?action=displaySalesOrder&ordheadID=${problem.linkedSalesOrderID}`,
+                    target: '_blank'
+                }, problem.linkedSalesOrderID)
+            },
+            {
                 path: "notes",
                 label: "",
                 key: "notes",